refactor(crdt): consolidate server ownership
