+++ apps/api-v1/test/routes/crdt-admin-repository-health.test.ts
@@ -10,7 +10,7 @@ import {
   type RallarCrdtDocumentRef,
   type RallarCrdtUpdateEnvelope,
 } from '@shared/crdt/mod.ts';
-import { InMemoryRallarCrdtLogRepository } from '@shared-server/crdt/InMemoryRallarCrdtLogRepository.ts';
+import { InMemoryRallarCrdtLogRepository } from '@shared-server/rallar-system/crdt/persistence/in-memory-crdt-log-repository.ts';
 
 import { registerCrdtAdminRoutes } from '../../src/routes/crdt-admin-routes.ts';
