+++ packages/shared-server/mod.ts
@@ -3,8 +3,10 @@ export * from './rallar-facade/RallarServerApplication.ts';
 export * from './rallar-facade/ws-topic-router.ts';
 export * from './game/mod.ts';
 export * from './rallar-ai/mod.ts';
-export * from './crdt/RallarCrdtServer.ts';
-export * from './crdt/InMemoryRallarCrdtLogRepository.ts';
+export * from './rallar-system/crdt/realtime/rallar-crdt-server-contracts.ts';
+export * from './rallar-system/crdt/realtime/install-rallar-crdt-ws-topics.ts';
+export * from './rallar-system/crdt/realtime/validate-rallar-crdt-server-live-envelope.ts';
+export * from './rallar-system/crdt/persistence/in-memory-crdt-log-repository.ts';
 export * from './app-data/AppDataRepository.ts';
 export * from './app-data/RallarServerAppData.ts';
 export * from './rallar-system/middleware/RallarMiddleware.ts';
