+++ packages/tests/shared-server/mutation-routing-inventory-decoding.ts
@@ -25,7 +25,7 @@ const PATHS = {
   l: 'packages/shared-server/rallar-system/services/ws-lifecycle-service.ts',
   e: 'packages/shared-server/rallar-system/group-state/presence/reconcile-expired-group-presence.ts',
   s: 'packages/shared-server/rallar-system/rtc-topology/topic/init-rtc-rtt-topic.ts',
-  d: 'packages/shared-server/crdt/RallarCrdtServer.ts',
+  d: 'packages/shared-server/rallar-system/crdt/realtime/install-rallar-crdt-ws-topics.ts',
   gi: 'packages/shared-server/rallar-system/services/AppGroupInboxService.ts',
   ci: 'apps/api-v1/src/composition/create-api-v1-system-installers.ts',
 } as const;
